+++ docs/docs/sandpack/plate-versions.ts
@@ -1,5 +1,5 @@
-export const plateVersion = '18.1.0';
+export const plateVersion = '18.8.1';
 export const testUtilsVersion = '13.4.0';
-export const excalidrawVersion = '17.0.3';
-export const juiceVersion = '17.0.3';
-export const selectionVersion = '17.0.3';
+export const excalidrawVersion = '18.7.0';
+export const juiceVersion = '18.7.0';
+export const selectionVersion = '18.7.0';