+++ examples/src/common/mapNodeId.ts
@@ -1,0 +1,7 @@
+let id = 0;
+
+export const mapNodeId = (nodes: any) =>
+  nodes.map((node: any) => {
+    id++;
+    return { ...node, id: id.toString() };
+  });