+++ packages/core/src/types/plugin/PlatePlugin.ts
@@ -1,5 +1,6 @@
 import { FC, ReactNode } from 'react';
 import { Value } from '../../slate/editor/TEditor';
+import { TEditableProps } from '../../slate/index';
 import { AnyObject } from '../misc/AnyObject';
 import { Nullable } from '../misc/Nullable';
 import { WithRequired } from '../misc/types';
@@ -155,14 +156,25 @@ export type PlatePlugin<
     }>;
 
     /**
+     * Render a component above `Slate`.
+     */
+    renderAboveSlate?: FC<{
+      children: ReactNode;
+    }>;
+
+    /**
      * Render a component after `Editable`.
      */
-    renderAfterEditable?: () => JSX.Element | null;
+    renderAfterEditable?: (
+      editableProps: TEditableProps<V>
+    ) => JSX.Element | null;
 
     /**
      * Render a component before `Editable`.
      */
-    renderBeforeEditable?: () => JSX.Element | null;
+    renderBeforeEditable?: (
+      editableProps: TEditableProps<V>
+    ) => JSX.Element | null;
 
     /**
      * Property used by `serializeHtml` util to replace `renderElement` and `renderLeaf` when serializing a node of this `type`.