+++ packages/nodes/link/src/components/FloatingLink/useFloatingLinkInsert.ts
@@ -28,7 +28,7 @@ export const useFloatingLinkInsert = ({
   const editor = useEditorRef();
   const focused = useFocused();
   const mode = useFloatingLinkSelectors().mode();
-  const open = useFloatingLinkSelectors().open();
+  const open = useFloatingLinkSelectors().isOpen(editor.id);
 
   const { triggerFloatingLinkHotkeys } = getPluginOptions<LinkPlugin>(
     editor,
@@ -37,7 +37,9 @@ export const useFloatingLinkInsert = ({
 
   useHotkeys(
     triggerFloatingLinkHotkeys!,
-    () => {
+    (e) => {
+      e.preventDefault();
+
       triggerFloatingLinkInsert(editor, {
         focused,
       });
@@ -48,14 +50,20 @@ export const useFloatingLinkInsert = ({
     [focused]
   );
 
-  const ref = useOnClickOutside(() => {
-    if (floatingLinkSelectors.mode() === 'insert') {
-      floatingLinkActions.hide();
-      focusEditor(editor, editor.selection!);
+  const ref = useOnClickOutside(
+    () => {
+      if (floatingLinkSelectors.mode() === 'insert') {
+        floatingLinkActions.hide();
+        focusEditor(editor, editor.selection!);
+      }
+    },
+    {
+      disabled: !open,
     }
-  });
+  );
 
   const { update, style, floating } = useVirtualFloatingLink({
+    editorId: editor.id,
     open: open && mode === 'insert',
     getBoundingClientRect: getSelectionBoundingClientRect,
     whileElementsMounted: () => {},