+++ packages/ui/dnd/src/dndStore.ts
@@ -1,0 +1,5 @@
+import { createStore } from '@udecode/plate-core';
+
+export const dndStore = createStore('dnd')({
+  isDragging: false,
+});