+++ packages/core/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @udecode/plate-core
 
+## 18.9.0
+
+### Minor Changes
+
+- [#1978](https://github.com/udecode/plate/pull/1978) by [@zbeyens](https://github.com/zbeyens) – Plugin fields `renderBeforeEditable` and `renderAfterEditable` now have `TEditableProps` passed as the first parameter.
+
+## 18.7.0
+
+### Minor Changes
+
+- [#1960](https://github.com/udecode/plate/pull/1960) by [@zbeyens](https://github.com/zbeyens) –
+  - Default editor value is now overridable with `editor.childrenFactory()`
+  - New core plugin `nodeFactory`, extends the editor with:
+    - `blockFactory: (node) => TElement`, can be used to create the default editor block
+    - `childrenFactory: () => Value`
+  - New transform `resetEditorChildren`: Replace editor children by `editor.childrenFactory()`.
+
+## 18.6.0
+
+### Minor Changes
+
+- [#1959](https://github.com/udecode/plate/pull/1959) by [@zbeyens](https://github.com/zbeyens) –
+  - Default editor value is now overridable with `editor.childrenFactory()`
+  - New core plugin `nodeFactory`, extends the editor with:
+    - `blockFactory: (node) => TElement`, can be used to create the default editor block
+    - `childrenFactory: () => Value`
+  - New transform `resetEditorChildren`: Replace editor children by `editor.childrenFactory()`.
+
+### Patch Changes
+
+- [#1957](https://github.com/udecode/plate/pull/1957) by [@tmilewski](https://github.com/tmilewski) – fix: update `@radix-ui/react-slot` to eliminate conflicting peer dependencies
+
+- [#1953](https://github.com/udecode/plate/pull/1953) by [@zbeyens](https://github.com/zbeyens) – `applyDeepToNodes`: new option `path`
+
+## 18.2.0
+
+### Minor Changes
+
+- [#1888](https://github.com/udecode/plate/pull/1888) by [@zbeyens](https://github.com/zbeyens) –
+  - new `PlatePlugin` property:
+    - `renderAboveSlate` – Render a component above `Slate`
+  - `id` is no longer required in plate hooks:
+    - `usePlateId()` is getting the closest editor id
+    - it's used in all store hooks if no store is found with the omitted id
+    - note that `id` is not needed if you don't have nested `Plate` / `PlateProvider`
+  - `id` prop change should remount `Plate`
+
 ## 18.1.1
 
 ### Patch Changes